aggiunta una delle funzioni per visualizzare il credito, cambiato l' attributo credito da double a float per motivi di conversione da postgres

+++ Backend/ProgettoAste2/src/main/java/com/gruppoaste2/progettoaste/model/UtenteRegistratoModel.java
@@ -5,9 +5,9 @@ import java.util.UUID;
 public class UtenteRegistratoModel extends UtenteModel {
 
     private final String numeroTelefono;
-    private final double credito;
+    private final float credito;
 
-    public UtenteRegistratoModel(UUID id, String username, String email, String numeroTelefono, String password, double credito){
+    public UtenteRegistratoModel(UUID id, String username, String email, String numeroTelefono, String password, float credito){
         super(id, username, email,password);
         this.numeroTelefono = numeroTelefono;
         this.credito = credito;
@@ -18,7 +18,7 @@ public class UtenteRegistratoModel extends UtenteModel {
         return numeroTelefono;
     }
 
-    public double getCredito()
+    public float getCredito()
     {
         return credito;
     }
